server/requirements.txt
# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
bcrypt==4.1.2

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging (built-in, but explicit for clarity)
# Python logging module is built-in

# Utilities
python-dotenv==1.0.0
numpy==1.26.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0